WIP: Change the interface to open image files
Add a struct xfig_stream that contains enough information to decide whether a
stream is from a regular file or a pipe, whether a file is compressed, how to
uncompress, and a possibility to obtain the uncompressed contents of the file.
Also, add functions to work with the above mentioned struct.