Hinder autoconf at replacing malloc() or realloc()
The AC_FUNC_MALLOC and AC_FUNC_REALLOC macros in configure.ac do a funny
check on malloc() and realloc(), namely, whether calling them with 0
returns a valid pointer. If the compile-switch -fsanitize=address is
used, malloc() and realloc() fail these tests, and autoconf defines
malloc and realloc to rpl_malloc and rpl_realloc. This causes the
replacement functions under lib/ to be linked in. But function
declarations are not provided for rpl_malloc() and rpl_realloc(), these
would have to be added (probably directly in config.h). So much hassle,
and then, who calls malloc() or realloc() with 0 and depends on them
returning a valid pointer when -fsanitize=address is in force? Remove
AC_FUNC_MALLOC and AC_FUNC_REALLOC.