Add a make target to build some documentation files
Enable the creation of some documentation files in the source tree with
"cd doc && make documentation". This will create
doc/html/images/xfig-title.png and doc/xfig_ref_en.pdf. The build
requires additional tools, viz., fig2dev and htmldoc.

The files mentioned above exist in the source tree, and it is not
necessary to re-create them. The make target is provided as a
convenience, for users who want to update the documentation.

This commit reduces the difference between files in the distribution
tarball and files in the development tree. Previously, it was thought
that users retrieve the sources mainly from the distribution tarball,
not from the source repository. Nowadays, there should not be a big
distinction between these two methods of publication.