Adding points to splines works if boxes are present
Create a figure containing a spline and a rectangle. Try to add a point
to the spline. With this commit, adding a point succeeds.

The function init_search() in u_search.c counts objects that may be a
target for editing, excluding some objects that do not pass a search
mask, see, e.g., validline_in_mask(). Next, these objects are searched
for points that are close enough to the cursor position, for instance
with the function next_line_found(). However, in these functions objects
that do not pass the search mask still were counted. Therefore, if
objects that are excluded in the first pass are present in the figure,
the loop terminated before all possible objects were checked. With this
commit, excluded objects do not contribute to the object count in either
function.