Fix a crash of scaling a figure with pattern fills
Changing the display zoom factor of a figure containing pattern fills
crashed the program. The field pattern_image[].cdata was initialized
with static data and it was tried to free() this data. Now, malloc() and
initialize pattern_images[].cdata properly in init_fill_pm(), which is
called once at the start of the program from setup_ind_panel().