Remove Sutherland-Hodgman polygon clipping
If CLIP_LINE was defined, lines and polygons were clipped to an
area larger than the canvas. Since CLIP_LINE was not noticed for
a long time, the entire clipping algorithm is removed.
Presumably, the graphics server does a better job at handling
graphics objects. When enlarging a drawing containing a polygon
with more than about 10 points, the clipping algorithm sometimes
caused a segfault.