Correct cancel zooming using a zoom area
When zooming in with the mouse, a zoom area is selected by
pressing Ctrl, clicking the primary mouse button to define one
corner of a rectangle and clicking again the primary mouse button
to define the second corner. The zoom area definition can be
cancelled by clicking, after the first mouse click, one of the
drawing modes in the left menu, or using an accelerator (pressing
l, b, e, or other keys). After cancelling in this way, the zoom
menu for the mouse buttons was not reachable any longer. This is
corrected.