Avoid crashing when re-editing text
When in text mode, clicking somewhere on the canvas to add a text
object, typing a few letters, clicking in the middle of the text,
typing Return (a new line opens), but clicking with the mouse
pointer again in the middle of the text would cause a segfault.
A similar segfault can be provoked by pressing undo. These
crashes are avoided with this commit.