Clear object cache after opening or closing a compound
Do not remember the state of the object cache over opening or
closing one or more compounds for individual editing. This led to
a double free by the sequence of opening a compound, adding a
line, closing this compound, pressing undo and adding a new line.
Also, as documented, deleting an object on the canvas, opening a
compound, and pressing immediately undo would undo the delete and
make the object part of the compound.

Now, pressing undo immediately after opening a compound would
close this compound. Pressing undo after closing one compound
would re-open this last compound. Pressing undo after closing all
compounds does nothing.

Perhaps this commit also fixes a change of positioning mode as
reported in ticket #51. However, I could not reproduce this issue.