Keep compound bound on editing/updating/aligning, #51
Previously, the size of a compound could change after editing,
updating or aligning elements in an compound. The size change
could occur regardless of whether anything was done in an edit or
update operation. For instance, create a box with linewidth 4, set
the point positioning mode to half an inch (xfig -startposnmode 4)
and put that box into a compound. After that, either clicking the
compound in edit mode, pressing done, or choosing update mode,
deselect all, clicking the compound would change the size of the
compound.

With this commit, the boundaries of a compound remain on the
current point positioning grid.

When editing a compound, changing e.g., the x-position of the top
left corner and pressing enter or clicking done, the elements of
the compound would shift by the change in the position, but the
bounds would be re-calculated again to the current point
positioning grid. This seems to be counter-intuitive.