WIP: write picture paths relative to cur_file_dir
Add functions external_path() and internal_path() to (i) construct
an internal representation of a file path, which might be either
an absolute path, a path relative to cur_file_dir, the directory
where the fig file is saved, or a path relative to the home
directory, and (ii) retrieve that internal representation and
convert it back. Picture file paths are used to compare pictures
for equality, hence the internal representations of picture file
paths should be unambiguous.