Rewind both a stream and the associated file descriptor
In the program, sometimes calls to fread() and read() are mixed.
Rewinding the file stream with rewind() does not necessarily set
the file descriptor to the beginning of the file. Likewise, setting
the file descriptor to the position 0 does not necessarily set the
file stream to the beginning. Therfore, both is done.