Accept empty grid value
Sometimes, in a locale with the decimal separator being different
from fullstop, on export a grid option of 0,000:0,000in was passed
to fig2dev, causing an error. Avoid this error.
Open a figure, print to batch, open another figure, again print
to batch: On the second printing "0,000" appears as grid spec.