Do not rely on linker capability allow-multiple-definition
Test binaries need symbols from xfig. These are provided by renaming
the main() function of xfig to xfigmain() and creating a convenience
library from all source files of xfig.

Previously, the main() function was not renamed and linking to the convenience
library only succeeded if the linker was capable of shadowing later definitions
of a symbol, here main() in the convenience library by main() in test binaries.
The test binaries were compiled with the option -Wl,--allow-multiple-definition
(equivalent to -Wl,-z,muldefs).