Remove option -boldFont, resource boldFont
The bold font was used to print the current file name at the top of the
xfig window, next to the "Help" button, and for messages in the popup
query window. For instance, create a file "exists.fig", copy it to
"new.fig", open new.fig and try to save it as "exists.fig". Then, a
popup window appears that displays the message:
    "exists.fig" already exists. Do you want to overwrite it?
This message was printed using the bold font.

It seems unnecessary to define a specific font for these two uses.

In addition, the bold font was nominally used as a label for the toggle
boxes in the zoom panel, next to "Integer zoom", and for the toggle
boxes in the arrow size panel. The zoom and arrow size panel are both
located at the bottom of the canvas. The toggle boxes next to "Gray" and
"Blank" in the layer side panel also had a label set in the bold font.
For all these toggle boxes the labels consisted of two blanks. The size
of the toggle boxes seemed to be independent of the label font, hence
the bold font was also removed here.