Specifically specify resource international
The Fig resource file contains the line "Fig*international: False" to
cancel any possible "*international: True" that might exist in the
resources. Explicitly set everywhere the international resource, either
to False or to the value chosen by the -international option or the
Fig.international resource.

If the "international" resource is specified, the X server queries the
locale of the application, chooses a font from the FontSet resource
appropriate to the locale and expects strings to be encoded in the
locale encoding. However, most application-provided strings are anyhow
ascii, e.g. "File", "Edit", "Save as..." etc. Hence, the international
resource can be usually set to False. This is now done specifically for
each widget at widget creation time. Previously, the application
resource file contained the line "Fig*international: False". That lead
to the same effect, but could be modified by the user.