Use international text by default
Xfig should now accept text entered in any locale and display this text
on the canvas as well as in the widgets, e.g., in the edit panel or as
a file name. The one exception is the command panel, where it was not
possible to correctly display characters beyond ascii.