Install an app-default file with commented out defaults
Since fallback X resources are defined from the the app-defaults
file in the distribution, it is not necessary to load the same
resources. When installing the app-defaults file, comment out these
resources. Thus a non-operating file is installed, but users might
get an idea of the resources used by xfig.