Install a valid app-default file, revert [0897b1]
X applications search for app-default files in specific locations,
they do not query the X-server. Hence, if a file with commented-out
defaults is found, the compiled-in fallback resources are not used.
Either install a valid app-defaults file, or do not install any
app-defaults file. The latter could be achieved by configuring
with --with-appdefaultdir=/dev/null.