Redraw characters selected in the character map panel
The character map panel consists of Command widgets. Clicking on one of
these widgets would invoke the notify() action, which calls the callback
registered with the widget. This callback than pastes the selected
character into the current text object. However, other default actions
registered with the Command widget would invert and redraw the
background of the widget. Thus, the glyph painted with xft onto the
drawable region would effectively be erased. Install a different set of
actions to keep the glyph painted with xft.