Paste characters using an event handler
Use an event handler instead of an action table and translations.
The code is simpler. This partially reverts the previous commit, [dd1bf6].
Also, redraw glyhs if the mouse hovers over it. When the character map
pops up, often the first few glpyhs are not drawn. Hovering with the
mouse over an empty cell makes the glyph visible.