Satisfy sed on macOS, ticket #162
Sed on mac does neither know alternation (\|) or \t for tab. Also, \n is
not recognized as newline in a replacement string. To insert a literal
newline in a make rule, define a shell variable containing a newline,
see info autoconf > Portable Make > Newlines in Make Rules. The
replacement of a shell variable only works in a double-quoted string,
where backslashes must also be further quoted. Therefore, split the
sed-script into a single-quoted and double-quoted part.